web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
docker==7.0.0
paho-mqtt==1.6.1
python-socketio==5.10.0
pyyaml==6.0.1
watchdog==3.0.0
filelock==3.13.1